Scipts/provisionalIndices.py: initial commit of script to fetch indices for realtime ops
Script fetches recent provisional and quicklook Kp, recent F10.7
and 45-day forecast of Ap (transformed to Kp) and F10.7. Data are written
to a test file.
To Do: Add an option parser to specify the target run directory to write the
index file to. This will require a modification to the reading of RamIndices.txt
to parse the dates in the file (this mod is pending).