Added an SCB call to initialization to remove gap
Previously none of the SCB variables were being calculated until after
a RAM had run for a bit. This made restarts and fresh runs behave
slightly differently, but was only noticable when the restart was
triggered on an SCB call (or on a crash). This change also will benefit
the SCE components once they are included. Because of this change it was
necessary to update the test files outputs.