Switched from PSpline and Slatec to GSL
Updated all files to use the new GSL routines instead of the previous interpolation, derivatives, and integration.
This change required an update to the test outputs as the numerics are slightly different.
Deleted all unused files and the entire srcSlatec/ folder.
Also added back the plasmasphere scripts that were left out in the previous versions.