Removed SWMF coupling from base modules and added gaussian filter
Removed all calls to ModRamCouple and ModScbCouple, which are used to couple RAM-SCB with the framework. This was done to make the sure the new coupling code that will be written doesn't conflict with anything.
Also added a gaussian moving filter to the SCB pressure to make sure SCB has an easier time with highly dynamic times.