Fixed a bug in minimum B
There was an issue during certain times when the minimum B would move
off equator. The code was previously handling this by simply always
using the minimum B value when calculating mirror points. This worked
for periods where there was still a single minima, that was just off
equator, but did not work when multiple minima appeared. This fix
address that issue although still makes the assumption that the minimum
B (one or more) is relatively close to the equator. This change required
an update to the test files.

Also included with this change is the ability to choose the number of
ghost cells used through a param, and an adjustment to the E field
indexing (0:nR vs 1:nR+1).