Updated share folder to current SWMF version share
In order to simplify the installation and building processes, updated
the share folder to the current SWMF version. This handles C code and
openmp by default. Updated Config.pl, schemeSetup.sh, and the base
Makefile to handle the changes (also had to update ModSceRun since one
of its dependancies changed names in the share folder)