Fixed issues with coupling to the SWMF
Moved all coupling code to src/ModRamCouple and srcInterface/IM_wrapper
and changed the method for calculation of the alpha and beta euler
potentials when coupling. These changes cause a slight slowdown when run
in coupled mode, but do not impact stand alone mode in any way.

Added document detailing new Euler potential method for use with the
SWMF.