Restructured external dependancies
Moved all external dependancies to their own folder which the RAM-SCB
compiles against. Extracted the CARPENTER subroutine and added it to
the ModRamPlasmasphere module so that it could continue to be used.
Future updates to the plasmasphere calculations will be confined to this
module.

Change to ModRamEField was because it was inheriting the pi_d type from
the weimer model instead of from RAM which it should not have been
doing.