Modified bounce averaging calculation
Update to the way the bounce averaged values are calculated. Previously
a new function would need to be written for each variable that needed to
be bounce averaged. Now a single function will handle all bounce
averaged variables.

The calling syntax for bounce averaging is;
GSL_BounceAverage(variable(nTheta), BMirror(nPa), thetaValue(nTheta),
BValue(nTheta), bounceAveragedVariable(nPa))

Note that by default the integration uses linear interpolation of the
variable so the number of points along the field line should be
sufficient to be approximated by a linear interpolation.

This update required an update to the test coefficients.