Added a new plasmasphere model
Added a plasmasphere model that evolves the cold electron population by
using ExB transportation combined with loss and refilling terms. In
order to differentiate between using the Carpenter and Anderson model vs
the new model a new parameter was created for the PARAM file
(#PLASMASPHERE). The format is;
    #PLASMASPHERE
    T            DoUsePlasmasphere (T/F)
    Transport    PlasmasphereModel (Carpenter/Transport)
    Rasmussen    RefillingTau      (Fixed/Analytic/Rasmussen)

The RefillingTau parameter can read from a file (Fixed), can be
calculated analytically using the method in DGCPM (Analytic), or can be
calculated using the Rasmussen model (Rasmussen). Note that for all
three versions the loss time scale is currently fixed. By default the
plasmasphere is turned off.

CAVEATS:
 - The new transport model has not been thoroughly tested and is in a
   beta state
 - The refilling rates being calculated by the models have only
   been loosely compared against observed refilling rates