Added a test for the EMIC and FLC options
Added testEMIC to the Makefile. A PARAM file and outputs to check
against were also generated. Date specific inputs (the AEindex.txt) were
added to the input folder. Universal inputs were added to the tar file.
These are then extracted and copied to the input_ram folder in the run
directory.