launch_run: builds rundir and launches run
e.g., python launch_run.py -o -d ~/foo bar
where foo is the target location for the run directory
and bar is the input file with run details like
start and end date. Currently bar is not used as the format
is TBD after consultation with CCMC as it will use whatever
is returned by the RoR web interface.